`(v2)/(authenticated)`を削除
Fixes #1065
