+++ app/(v2)/NotifyVerification.tsx
