+++ app/(v2)/home/TimelineSegment.stories.tsx
