+++ app/(v2)/home/TimelineSegment.tsx
