+++ app/(v2)/layout.tsx
@@ -8,7 +8,7 @@ import GlobalFooter from "~/components/GlobalFooter";
 import GlobalNav from "~/components/GlovalNav";
 import { ToastProvider } from "~/components/Toaster";
 
-import { NotifyEmailVerifycation } from "./(authenticated)/home/NotifyVerification";
+import { NotifyEmailVerifycation } from "./NotifyVerification";
 
 export default function ApplicationLayout({
   children,
