+++ app/(v2)/users/[name]/SideNav.tsx
@@ -1,8 +1,8 @@
 import clsx from "clsx";
 import { ReactNode } from "react";
 
-import MyLikesPageLink from "~/app/(v2)/(authenticated)/me/likes/Link";
-import MyMylistsPageLink from "~/app/(v2)/(authenticated)/me/mylists/Link";
+import MyLikesPageLink from "~/app/(v2)/me/likes/Link";
+import MyMylistsPageLink from "~/app/(v2)/me/mylists/Link";
 import UserLikesPageLink2 from "~/app/(v2)/users/[name]/likes/Link";
 import UserMylistsPageLink from "~/app/(v2)/users/[name]/mylists/Link";
 import { FragmentType, graphql, useFragment } from "~/gql";
