+++ components/GlovalNav/ProfileAccordion.tsx
@@ -3,10 +3,10 @@
 import clsx from "clsx";
 import React from "react";
 
-import MyLikesPageLink from "~/app/(v2)/(authenticated)/me/likes/Link";
-import MyTopPageLink from "~/app/(v2)/(authenticated)/me/Link";
-import MyMylistsPageLink from "~/app/(v2)/(authenticated)/me/mylists/Link";
-import NotificationsPageLink from "~/app/(v2)/(authenticated)/notifications/Link";
+import MyLikesPageLink from "~/app/(v2)/me/likes/Link";
+import MyTopPageLink from "~/app/(v2)/me/Link";
+import MyMylistsPageLink from "~/app/(v2)/me/mylists/Link";
+import NotificationsPageLink from "~/app/(v2)/notifications/Link";
 import SettingPageLink from "~/app/(v2)/settings/Link";
 import { LogoutLink } from "~/components/AuthLink";
 import { FragmentType, graphql, useFragment } from "~/gql";
